correction on intarray.C to allow compilation with MSVC

+++ src/oofemlib/intarray.C
@@ -257,7 +257,7 @@ void IntArray :: printYourself(const std::string name) const
 
 bool IntArray :: isFinite() const
 {
-    for(int val : values) {
+	for (double val : values) {
         if(!std::isfinite(val)) {
             return false;
         }
